explicit type conversion

+++ homework_2/hw2.4/main.c
@@ -24,7 +24,7 @@ void inputNumerals(int* numerals, int n)
     char* number = (char*)calloc(n + 1, sizeof(char));
     scanf("%s", number);
     for (int i = n - 1; i >= 0; --i) {
-        numerals[i] = number[i] - '0';
+        numerals[i] = (int)(number[i] - '0');
     }
     free(number);
 }
